feat: add rating service and docker

+++ components/profile_service/models.py
@@ -1,5 +1,7 @@
-from sqlalchemy import Column, Integer, String, Text
+from sqlalchemy import Column, BigInteger, String, Text, Integer
 from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import relationship
+
 
 Base = declarative_base()
 
@@ -7,7 +9,7 @@ Base = declarative_base()
 class Profile(Base):
     __tablename__ = "profiles"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     first_name = Column(String(100), nullable=False)
     last_name = Column(String(100), nullable=False)
     bio = Column(Text, nullable=True)
@@ -15,3 +17,5 @@ class Profile(Base):
     gender = Column(String(10), nullable=False)  # Male, Female, Other
     city = Column(String(100), nullable=False)
     photo_path = Column(String(255), nullable=True)  # To save the photo file path
+
+    # rating = relationship("ProfileRating", uselist=False, back_populates="profile")  # One-to-one relationship
